requirements.txt
requests
psycopg2-binary
beautifulsoup4
alembic==1.6.5
Faker==8.5.1
Flask==2.0.1
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1